+++ src/layouts/TabLayout/TabLayout.styles.ts
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+	container: {
+		paddingHorizontal: 24
+	},
+	tabsContainer: {
+		flex: 1,
+		backgroundColor: '#fff',
+		padding: 10
+	}
+});
