+++ src/layouts/TabLayout/TabLayout.types.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export interface TabLayoutProps {
+	left: React.ReactChild;
+	right: React.ReactChild;
+	leftTitle: string;
+	rightTitle: string;
+}
