+++ src/layouts/TabLayout/TabSelector/Tab/Tab.types.ts
@@ -1,0 +1,7 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface TabProps {
+	active?: boolean;
+	side: string;
+	onTabSelect: Dispatch<SetStateAction<string>>;
+}
