+++ src/layouts/TabLayout/TabSelector/TabSelector.tsx
@@ -1,32 +1,15 @@
-import React, { Dispatch, SetStateAction } from 'react';
-import { View, StyleSheet, Dimensions } from 'react-native';
+import React from 'react';
+import { View } from 'react-native';
 import { Text } from '@components';
 import { useTheme } from '@hooks';
-import Tab from './Tab';
-
-const screenWidth = Dimensions.get('window').width;
-
-const styles = StyleSheet.create({
-	tabContainer: {
-		height: 51,
-		flexDirection: 'row',
-		position: 'absolute',
-		bottom: 0,
-		width: screenWidth
-	}
-});
-
-interface TabSelectorProps {
-	setSelectedTab: Dispatch<SetStateAction<string>>;
-	selectedTab: string;
-	leftTitle: string;
-	rightTitle: string;
-}
+import Tab from './Tab/Tab';
+import styles from './TabSelector.styles';
+import { TabSelectorProps } from './TabSelector.types';
 
 const TabSelector: React.FC<TabSelectorProps> = ({ setSelectedTab, selectedTab, leftTitle, rightTitle }) => {
 	const { colors } = useTheme();
 	return (
-		<View style={[styles.tabContainer, { backgroundColor: colors.background1 }]}>
+		<View style={[styles.container, { backgroundColor: colors.background1 }]}>
 			<Tab side="left" active={selectedTab === 'transactions' && true} onTabSelect={setSelectedTab}>
 				<Text style={{ fontSize: 16 }} weight={selectedTab === 'transactions' ? 'bold' : 'regular'}>
 					{leftTitle}
