+++ src/layouts/TabLayout/TabSelector/TabSelector.types.ts
@@ -1,0 +1,8 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface TabSelectorProps {
+	setSelectedTab: Dispatch<SetStateAction<string>>;
+	selectedTab: string;
+	leftTitle: string;
+	rightTitle: string;
+}
