+++ src/screens/WalletScreen/WalletScreen.hooks.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import React, { useCallback, useEffect } from 'react';
 import { Alert } from 'react-native';
 import { useFocusEffect } from '@react-navigation/native';
